put overview logic into service #16
